Meldungsposition nicht mehr außerhalb des Zuständigkeitsbereichs setzbar (Bearbeiter). Schließen von Meldungen außerhalb des Zuständigkeitsbereiches
