+++ app/models/issue/callbacks.rb
@@ -6,6 +6,7 @@ class Issue
 
     include AuthorBlacklist
     include ConfirmationWithHash
+    include Setters
 
     included do
       before_destroy :destroy_dangling_associations
@@ -28,6 +29,7 @@ class Issue
                 saved_change_to_group_id? && !status_pending?
             }
 
+      validate :issue_in_districts, on: :update
       validates :description, :position, :status, presence: true
       validates :status_note, length: { maximum: Settings::Issue.status_note_max_length }
       validates :status_note, presence: true, if: :expected_closure_changed?
@@ -72,55 +74,6 @@ class Issue
       self.property_owner = Geocodr.property_owner(self)
     end
 
-    def reset_archived
-      self.archived_at = nil
-    end
-
-    def set_responsibility
-      return if group.present? && responsibility_action.blank?
-      return self.responsibility_accepted = true if responsibility_action_accept?
-      return self.responsibility_accepted = false if responsibility_action_reject?
-      recalculate_responsibility if recalculate_responsibility?
-      self.responsibility_accepted = group_id == group_id_was
-    end
-
-    def recalculate_responsibility?
-      responsibility_action_recalculate? || responsibility_action.nil?
-    end
-
-    def recalculate_responsibility
-      self.group = category&.group(lat: lat, lon: lon) || group
-    end
-
-    def responsibility_action_accept?
-      responsibility_action&.to_sym == :accept
-    end
-
-    def responsibility_action_recalculate?
-      responsibility_action&.to_sym == :recalculate
-    end
-
-    def responsibility_action_reject?
-      responsibility_action&.to_sym == :reject
-    end
-
-    def set_reviewed_at
-      self.reviewed_at = Time.current
-    end
-
-    def set_expected_closure
-      self.expected_closure = status_in_process? ? Time.zone.today + category.average_turnaround_time.days : nil
-    end
-
-    def set_trust_level
-      self.trust_level = calculate_trust_level
-    end
-
-    def set_updated_by
-      self.updated_by_user = Current.user
-      self.updated_by_auth_code = Current.user&.auth_code
-    end
-
     def calculate_trust_level
       return 0 if (user = User.find_by(User.arel_table[:email].matches(author))).blank?
       return 2 if user.groups.any?(&:kind_field_service_team?)
@@ -146,6 +99,16 @@ class Issue
 
     def clear_group_responsibility_notified_at
       self.group_responsibility_notified_at = nil
+    end
+
+    def issue_in_districts
+      return if Current.user.blank? || Current.user.role_admin?
+      errors.add(:position, :outside_of_designated_districts) unless position_within_districts?
+    end
+
+    def position_within_districts?
+      return false unless Current.user
+      Current.user.districts.find_by('ST_Within(?, area)', position).present?
     end
   end
 end
