+++ config/locales/de.yml
@@ -137,6 +137,7 @@ de:
               blank: Die ausgewählte Kategorie wird derzeit an dieser Position nicht unterstützt.
             position:
               outside_instance: befindet sich außerhalb des gültigen Bereichs.
+              outside_of_designated_districts: befindet sich außerhalb Ihres Zuständigkeitsbereichs.
         responsibility:
           attributes:
             base:
@@ -423,6 +424,8 @@ de:
     responsibility_actions:
       current_responsibility: "aktuelle Zuständigkeit:"
       responsibility_action_accept: Zuständigkeit für die Meldung akzeptieren
+      responsibility_action_close_as_not_solvable:
+        Meldung befindet sich außerhalb des Zuständigkeitsbereichs (als nicht lösbar schließen)
       responsibility_action_manual: Zuständigkeit manuell setzen
       responsibility_action_recalculation: Zuständigkeit automatisch neu zuweisen
       responsibility_action_recalculation_review: "automatisch zugewiesene Zuständigkeit akzeptieren:"
