+++ config/status_note_template.sample.yml
@@ -5,3 +5,4 @@ Lichtraumprofil: Die Einschränkung der Durchgangsbreite ist noch nicht wesentli
 Pflegeerfordernis: Ihr Hinweis ist berechtigt. Er wird in die turnusmäßige Pflege/Säuberung eingeordnet.
 Pflegezustand: Ihr Hinweis ist nicht berechtigt. Der Pflegezustand/Reinigungszustand entspricht der Pflegestandard-Einstufung.
 Unmut: Ihr Hinweis zielt auf übergeordnete kommunalpolitische Belange ab. Eine Bearbeitung ist uns mittels dieses Mediums nicht möglich.
+Zuständigkeit: Widerspruch zwischen Meldungsposition und Inhalt der Meldung. Bitte erstellen Sie eine neue Meldung an der richtigen Position.
